added best first search 3d algorithm
