+++ src/com/gubertmc/plugin/main/algorithms/astar/astar2d/AstarSimulation2D.java
@@ -1,14 +1,14 @@
-package com.gubertmc.plugin.main.algorithms.bestfirstsearch.bestfirstsearch2d;
+package com.gubertmc.plugin.main.algorithms.astar.astar2d;
 
-import com.gubertmc.plugin.main.algorithms.Node;
 import com.gubertmc.plugin.main.algorithms.Simulation;
+import com.gubertmc.plugin.main.algorithms.Node;
 
 import java.util.ArrayList;
 import java.util.PriorityQueue;
 
-public class BestFirstSearchSimulation2D extends Simulation {
+public class AstarSimulation2D extends Simulation {
 
-    public BestFirstSearchSimulation2D(int[][][] maze, int[] startCoordinate, int[] endCoordinate) {
+    public AstarSimulation2D(int[][][] maze, int[] startCoordinate, int[] endCoordinate) {
         super(maze, startCoordinate, endCoordinate, false);
     }
 
@@ -37,9 +37,11 @@ public class BestFirstSearchSimulation2D extends Simulation {
         }
 
         Node currentNode = getCurrentNode();
+        int g = calculateG(currentNode);
+        currentNode.setG(g);
         int h = calculateH(currentNode);
         currentNode.setH(h);
-        currentNode.setBfsF();
+        currentNode.setF();
         setCurrentNode(currentNode);
         setStartNode(currentNode);
         PriorityQueue<Node> openList = getOpenList();
@@ -101,6 +103,34 @@ public class BestFirstSearchSimulation2D extends Simulation {
         return pathFound;
     }
 
+    public int calculateG(Node node) {
+        int row = node.getRow();
+        int col = node.getCol();
+        if (row == getCurrentNode().getRow() && col == getCurrentNode().getCol()) {
+            return 0;
+        }
+
+        Node parent = node.getParent();
+        if (parent == null) {
+            int xDistance;
+            if (col > getCurrentNode().getCol()) {
+                xDistance = col - getCurrentNode().getCol();
+            } else {
+                xDistance = getCurrentNode().getCol() - col;
+            }
+
+            int yDistance;
+            if (row > getCurrentNode().getRow()) {
+                yDistance = row - getCurrentNode().getRow();
+            } else {
+                yDistance = getCurrentNode().getRow() - row;
+            }
+
+            return (xDistance * 10) + (yDistance * 10);
+        }
+        return 10 + parent.getG();
+    }
+
     public int calculateH(Node node) {
         int row = node.getRow();
         int col = node.getCol();
@@ -139,9 +169,11 @@ public class BestFirstSearchSimulation2D extends Simulation {
                 && !getClosedList().contains(getGrid()[row - 1][col][zNum])) {
             Node[][][] grid = getGrid();
             grid[row - 1][col][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row - 1][col][zNum]);
+            grid[row - 1][col][zNum].setG(g);
             int h = calculateH(grid[row - 1][col][zNum]);
             grid[row - 1][col][zNum].setH(h);
-            grid[row - 1][col][zNum].setBfsF();
+            grid[row - 1][col][zNum].setF();
             setGrid(grid);
             PriorityQueue<Node> openList = getOpenList();
             openList.add(grid[row - 1][col][zNum]);
@@ -156,9 +188,11 @@ public class BestFirstSearchSimulation2D extends Simulation {
                 && !getClosedList().contains(getGrid()[row][col + 1][zNum])) {
             Node[][][] grid = getGrid();
             grid[row][col + 1][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row][col + 1][zNum]);
+            grid[row][col + 1][zNum].setG(g);
             int h = calculateH(grid[row][col + 1][zNum]);
             grid[row][col + 1][zNum].setH(h);
-            grid[row][col + 1][zNum].setBfsF();
+            grid[row][col + 1][zNum].setF();
             setGrid(grid);
             PriorityQueue<Node> openList = getOpenList();
             openList.add(grid[row][col + 1][zNum]);
@@ -173,9 +207,11 @@ public class BestFirstSearchSimulation2D extends Simulation {
                 && !getClosedList().contains(getGrid()[row + 1][col][zNum])) {
             Node[][][] grid = getGrid();
             grid[row + 1][col][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row + 1][col][zNum]);
+            grid[row + 1][col][zNum].setG(g);
             int h = calculateH(grid[row + 1][col][zNum]);
             grid[row + 1][col][zNum].setH(h);
-            grid[row + 1][col][zNum].setBfsF();
+            grid[row + 1][col][zNum].setF();
             setGrid(grid);
             PriorityQueue<Node> openList = getOpenList();
             openList.add(grid[row + 1][col][zNum]);
@@ -190,9 +226,11 @@ public class BestFirstSearchSimulation2D extends Simulation {
                 && !getClosedList().contains(getGrid()[row][col - 1][zNum])) {
             Node[][][] grid = getGrid();
             grid[row][col - 1][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row][col - 1][zNum]);
+            grid[row][col - 1][zNum].setG(g);
             int h = calculateH(grid[row][col - 1][zNum]);
             grid[row][col - 1][zNum].setH(h);
-            grid[row][col - 1][zNum].setBfsF();
+            grid[row][col - 1][zNum].setF();
             setGrid(grid);
             PriorityQueue<Node> openList = getOpenList();
             openList.add(grid[row][col - 1][zNum]);
