+++ src/com/gubertmc/plugin/main/mazes/custom/AstarMaze3D.java
@@ -3,9 +3,9 @@ package com.gubertmc.plugin.main.mazes.custom;
 import com.gubertmc.MazeGeneratorPlugin;
 import com.gubertmc.plugin.main.algorithms.Animation;
 import com.gubertmc.plugin.main.algorithms.Simulation;
-import com.gubertmc.plugin.main.algorithms.astar.astar2d.PathfindingAnimation2D;
-import com.gubertmc.plugin.main.algorithms.astar.astar2d.PathfindingSimulation2D;
-import com.gubertmc.plugin.main.mazes.Maze2D;
+import com.gubertmc.plugin.main.algorithms.astar.astar3d.AstarAnimation3D;
+import com.gubertmc.plugin.main.algorithms.astar.astar3d.AstarSimulation3D;
+import com.gubertmc.plugin.main.mazes.Maze3D;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
@@ -13,9 +13,9 @@ import org.bukkit.scheduler.BukkitRunnable;
 
 import static org.bukkit.Bukkit.getServer;
 
-public class PathfindingMaze2D extends Maze2D {
+public class AstarMaze3D extends Maze3D {
 
-    public PathfindingMaze2D(MazeGeneratorPlugin plugin, Block block, int size, double wallPercentage) {
+    public AstarMaze3D(MazeGeneratorPlugin plugin, Block block, int size, double wallPercentage) {
         super(plugin, block, size, wallPercentage);
     }
 
@@ -43,7 +43,7 @@ public class PathfindingMaze2D extends Maze2D {
         int count = 0;
         while (!isValid()) {
             int[][][] simulationMaze = generateSimulation();
-            Simulation simulation = new PathfindingSimulation2D(
+            Simulation simulation = new AstarSimulation3D(
                     simulationMaze, getStartCoordinate(), getEndCoordinate()
             );
             simulation.setup();
@@ -53,7 +53,6 @@ public class PathfindingMaze2D extends Maze2D {
                 setTime(0L);
                 generateCore(coreMaterial, adjustedTime);
                 // generateBorder(coreMaterial);
-                clearOldBeacons();
                 generateStartAndEndPoints(startPointGlassMaterial, endPointGlassMaterial);
                 generateBlockedAreas(simulationMaze, blockerMaterial);
 
@@ -63,7 +62,7 @@ public class PathfindingMaze2D extends Maze2D {
                 new BukkitRunnable() {
                     @Override
                     public void run() {
-                        Animation animation = new PathfindingAnimation2D(
+                        Animation animation = new AstarAnimation3D(
                                 getPlugin(),
                                 getMazeBlockLocations(),
                                 getStartCoordinate(),
@@ -78,8 +77,8 @@ public class PathfindingMaze2D extends Maze2D {
                         );
                         animation.setup();
                         setValid(animation.start());
-                        getServer().broadcastMessage(ChatColor.GREEN + "" + getSize() + "x" + getSize()
-                                + " maze generated..."
+                        getServer().broadcastMessage(ChatColor.GREEN + "" + getSize() + "x" + getSize() + "x"
+                                + getSize() + " maze generated..."
                         );
                         animation.showAnimation(getTime());
                         cancel();
@@ -92,9 +91,9 @@ public class PathfindingMaze2D extends Maze2D {
                 System.out.println("Invalid maze - starting new simulation...");
             }
             if (count == 50) {
-                getServer().broadcastMessage(ChatColor.RED + "A maze could not be successfully generated within " +
-                        "50 simulations. You may experience server lag. Creating a larger maze with a lower " +
-                        "percentage of walls/blockers will greatly help and put less stress on the server."
+                getServer().broadcastMessage(ChatColor.RED + "A maze could not be successfully generated within" +
+                        " 50 simulations. You may experience server lag. Creating a larger maze with a lower" +
+                        " percentage of walls/blockers will greatly help and put less stress on the server."
                 );
             }
         }
